chore: move adapter to limits.go
